LayoutTests/crypto/workers/subtle/rsa-postMessage-worker-expected.txt

Test sending rsa crypto keys via postMessage to a worker.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS All checks passed in worker for public key
PASS publicKey.type is 'public'
PASS publicKey.extractable is true
PASS publicKey.algorithm.name is 'RSAES-PKCS1-v1_5'
PASS publicKey.algorithm.modulusLength is 2048
PASS bytesToHexString(publicKey.algorithm.publicExponent) is '010001'
PASS publicKey.algorithm.hash is undefined.
PASS publicKey.usages is ['encrypt']
PASS exportedKey.n is publicKeyJSON.n
PASS exportedKey.e is publicKeyJSON.e
PASS All checks passed in worker for private key
PASS privateKey.type is 'private'
PASS privateKey.extractable is true
PASS privateKey.algorithm.name is 'RSAES-PKCS1-v1_5'
PASS privateKey.algorithm.modulusLength is 2048
PASS bytesToHexString(privateKey.algorithm.publicExponent) is '010001'
PASS privateKey.algorithm.hash is undefined.
PASS privateKey.usages is ['decrypt']
PASS exportedKey.n is privateKeyJSON.n
PASS exportedKey.e is privateKeyJSON.e
FAIL exportedKey.d should be eNLS37aCz7RXSNPD_DtLBJ6j5T8cSxdzRBCjPaI6WcGqJp16lq3UTwuoDLAqlA9oGYm238dsIWpuucP_lQtbWe-7SpxoI6_vmYGf7YVUHv1-DF9qiOmSrMmdxMnVOzYXY8RaT6thPjn_J5cfLV2xI_LwsrMtmpdSyNlgX0zTUhwtuahgAKMEChYjH2EnjHdHw6sY2-wApdcQI7ULE0oo5RzbQZpmuhcN9hiBc0L3hhF0qo50mbl02_65_GQ7DpVkXBxNgRBLzlPabmzzG2oAhfefLgYmSC1opaCkXE6vRWQNWNL45RZNZFYM3uoJghOMqGeocM0BpjdChHrPOlFvSQ. Was IfKI2jcblgtsk50iazLxL3m6IovELcfrLOkWeAtlbiFw4BzZYZDRiXXn34BhqVx7ZIqeGl4DSpbkvu_qrsTs3YxVpmNk4v3se8rtYHRAuTGHWG7sUrF9iSrEWPybs3J7X59C2PEYKBz2WQiXH5QbHOpzWP8in2FO_3uuWb8tJTTPqyi1z3n04uNWBGjnYfUAeMOXyq4GSzAcCO55FEZeFQpuNIQBAaC1NCqPps4ps2AEoMBC6Kdg0AI8xwPGysq_qnFbNlkCqopRKiu3bs5pnOzOHpQCb8fYNeIih4qPsKaPQ0lHwFAAo-AsE_7niiPSJQwzB-dcoVfSVXss6TZ49Q.
PASS exportedKey.p is privateKeyJSON.p
PASS exportedKey.q is privateKeyJSON.q
PASS exportedKey.dp is privateKeyJSON.dp
PASS exportedKey.dq is privateKeyJSON.dq
PASS exportedKey.qi is privateKeyJSON.qi
PASS successfullyParsed is true

TEST COMPLETE
